+++ src/main/java/com/hexvane/aetherhaven/command/AetherhavenJournalCommand.java
@@ -18,6 +18,11 @@ public final class AetherhavenJournalCommand extends AbstractPlayerCommand {
     }
 
     @Override
+    protected boolean canGeneratePermission() {
+        return false;
+    }
+
+    @Override
     protected void execute(
         @Nonnull CommandContext context,
         @Nonnull Store<EntityStore> store,